+++ app/src/main/AndroidManifest.xml
@@ -20,10 +20,10 @@
         android:theme="@style/Theme.FromPet"
         tools:targetApi="31">
         <activity
-            android:name=".ui.chat.ChatClickUserDetailActivity"
+            android:name=".ui.chat.activity.ChatClickUserDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.chat.ChatPullScreenActivity"
+            android:name=".ui.chat.activity.ChatPullScreenActivity"
             android:exported="false" />
 
         <meta-data
@@ -34,10 +34,10 @@
             android:name=".ui.setting.SettingProfileActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.chat.ChatUserDetailActivity"
+            android:name=".ui.chat.activity.ChatUserDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.chat.ChatMessageActivity"
+            android:name=".ui.chat.activity.ChatMessageActivity"
             android:exported="false"
             android:windowSoftInputMode="adjustResize" />
         <activity
