+++ app/src/main/java/com/example/frompet/data/repository/matched/MatchedRepositoryImpl.kt
@@ -1,0 +1,4 @@
+package com.example.frompet.data.repository.matched
+
+class MatchedRepositoryImpl {
+}
