+++ app/src/main/java/com/example/frompet/ui/chat/fragment/ChatFragment.kt
@@ -1,4 +1,4 @@
-package com.example.frompet.ui.chat
+package com.example.frompet.ui.chat.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
