+++ app/src/main/res/layout/fragment_chat.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.chat.ChatFragment"
+    tools:context=".ui.chat.fragment.ChatFragment"
     android:background="@color/light_gray">
 
         <androidx.viewpager2.widget.ViewPager2
