+++ app/src/main/res/layout/fragment_chat_list.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/light_gray"
     app:layoutDescription="@xml/fragment_chat_list_scene"
-    tools:context=".ui.chat.ChatLikeListFragment">
+    tools:context=".ui.chat.fragment.ChatLikeListFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/constraintLayout3"
