+++ app/src/main/res/layout/item_chatlist.xml
@@ -8,7 +8,7 @@
     app:cardUseCompatPadding="true"
     app:cardCornerRadius="13dp"
     app:cardElevation="10dp"
-    tools:context=".ui.chat.ChatLikeListFragment">
+    tools:context=".ui.chat.fragment.ChatLikeListFragment">
     <LinearLayout
         android:orientation="vertical"
         android:padding="13dp"
