+++ app/src/main/res/values/colors.xml
@@ -16,6 +16,7 @@
     <color name="light_orenge">#F68531</color>
     <color name="light_gray">#F4F5F9</color>
     <color name="background">#FEE48D</color>
+    <color name="red">#F44040</color>
